requirements.txt
flask==2.0.3
beautifulsoup4==4.9.3
requests==2.25.1
python-dotenv==0.19.0
werkzeug==2.0.3